rename class and reorganize files
